+++ CHANGELOG.md
@@ -1,3 +1,8 @@
 # zodme Changelog
 
 ## [Initial Version] - {PR_MERGE_DATE}
+
+## First Update - 2026-07-26
+
+- New command incorporated: Browse All Zodiac Signs
+- Zodiac Sign Description Enrichment